Move Create and Extend... traits to separate submodule
